add --no-backup option

+++ lib/aws_rotate/backup.rb
@@ -1,7 +1,7 @@
 module AwsRotate
   class Backup < Base
     def run
-      return if @options[:noop]
+      return if @options[:noop] || @options[:backup] == false
       return unless credentials_exist?
       backup_path = @credentials_path + ".bak-#{Time.now.strftime("%F-%T")}"
       FileUtils.cp(@credentials_path, backup_path)
